# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                },
                "example": [
                  {
                    "id": "67efcab4b388d42bb5bf6286",
                    "email": "email@gmail.com",
                    "password": "123456",
                    "createdAt": "2025-04-04T12:56:19.903Z",
                    "updatedAt": "2025-04-04T12:56:19.903Z",
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InN0cmluZyIsImlkIjoiNjdmMzhkMmNmYmM3MjBiZjM1MDIyYzdlIiwicm9sZXMiOlsiVVNFUiJdLCJpYXQiOjE3NDQwMTUyMzcsImV4cCI6MTc0NDEwMTYzN30.eEV89O1ZRMlGEYE7CH4JGhTpR1FCMzfGrFvXl9xKVeI"
                  }
                ]
              }
            }
          }
        },
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get one user by id",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/users/{email}": {
      "get": {
        "operationId": "UsersController_getUserByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                },
                "example": [
                  {
                    "id": "67efcab4b388d42bb5bf6286",
                    "email": "email@gmail.com",
                    "password": "123456",
                    "createdAt": "2025-04-04T12:56:19.903Z",
                    "updatedAt": "2025-04-04T12:56:19.903Z",
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InN0cmluZyIsImlkIjoiNjdmMzhkMmNmYmM3MjBiZjM1MDIyYzdlIiwicm9sZXMiOlsiVVNFUiJdLCJpYXQiOjE3NDQwMTUyMzcsImV4cCI6MTc0NDEwMTYzN30.eEV89O1ZRMlGEYE7CH4JGhTpR1FCMzfGrFvXl9xKVeI"
                  }
                ]
              }
            }
          }
        },
        "summary": "Get one user by email",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/roles": {
      "post": {
        "operationId": "RolesController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                },
                "example": {
                  "_id": "67efcab4b388d42bb5bf6286",
                  "value": "ADMIN",
                  "createdAt": "2025-04-04T12:56:19.903Z",
                  "updatedAt": "2025-04-04T12:56:19.903Z"
                }
              }
            }
          }
        },
        "summary": "Create role",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "RolesController_getAllRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleDto"
                  }
                },
                "example": [
                  {
                    "_id": "67efcab4b388d42bb5bf6286",
                    "value": "ADMIN",
                    "createdAt": "2025-04-04T12:56:19.903Z",
                    "updatedAt": "2025-04-04T12:56:19.903Z"
                  }
                ]
              }
            }
          }
        },
        "summary": "Get all roles",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/roles/{role}": {
      "get": {
        "operationId": "RolesController_getRoleByValue",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                },
                "example": {
                  "_id": "67efcab4b388d42bb5bf6286",
                  "value": "ADMIN",
                  "createdAt": "2025-04-04T12:56:19.903Z",
                  "updatedAt": "2025-04-04T12:56:19.903Z"
                }
              }
            }
          }
        },
        "summary": "Get role",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                },
                "example": {
                  "id": "67efcab4b388d42bb5bf6286",
                  "email": "email@gmail.com",
                  "password": "123456",
                  "createdAt": "2025-04-04T12:56:19.903Z",
                  "updatedAt": "2025-04-04T12:56:19.903Z",
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InN0cmluZyIsImlkIjoiNjdmMzhkMmNmYmM3MjBiZjM1MDIyYzdlIiwicm9sZXMiOlsiVVNFUiJdLCJpYXQiOjE3NDQwMTUyMzcsImV4cCI6MTc0NDEwMTYzN30.eEV89O1ZRMlGEYE7CH4JGhTpR1FCMzfGrFvXl9xKVeI"
                }
              }
            }
          }
        },
        "tags": [
          "Authorization"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/auth/registration": {
      "post": {
        "operationId": "AuthController_registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                },
                "example": {
                  "id": "67efcab4b388d42bb5bf6286",
                  "email": "email@gmail.com",
                  "password": "123456",
                  "createdAt": "2025-04-04T12:56:19.903Z",
                  "updatedAt": "2025-04-04T12:56:19.903Z",
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InN0cmluZyIsImlkIjoiNjdmMzhkMmNmYmM3MjBiZjM1MDIyYzdlIiwicm9sZXMiOlsiVVNFUiJdLCJpYXQiOjE3NDQwMTUyMzcsImV4cCI6MTc0NDEwMTYzN30.eEV89O1ZRMlGEYE7CH4JGhTpR1FCMzfGrFvXl9xKVeI"
                }
              }
            }
          }
        },
        "tags": [
          "Authorization"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "Generated Swagger JSON",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "uniqueItems": true
          },
          "email": {
            "type": "string",
            "uniqueItems": true
          },
          "password": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "password",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "MANAGER"
            ]
          }
        },
        "required": [
          "value"
        ]
      },
      "RoleDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "object",
            "uniqueItems": true
          },
          "value": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "MANAGER"
            ]
          }
        },
        "required": [
          "_id",
          "value"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "uniqueItems": true
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}